Fix json Views y mejoras
