+++ app/controllers/TvShowRequestController.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.google.inject.Inject;
-import utils.json.TvShowViews;
+import utils.json.JsonViews;
 import models.TvShow;
 import models.TvShowRequest;
 import models.service.SeasonService;
@@ -229,7 +229,7 @@ public class TvShowRequestController extends Controller {
 
                     // respuesta ok - devolvemos datos obtenidos
                     try {
-                      JsonNode jsonNode = jsonUtils.jsonParseObject(tvShow, TvShowViews.FullTvShow.class);
+                      JsonNode jsonNode = jsonUtils.jsonParseObject(tvShow, JsonViews.FullTvShow.class);
                       result.set("tvShow", jsonNode);
                       return ok(result);
                     } catch (JsonProcessingException e) {
