+++ app/models/TvShowRequest.java
@@ -1,7 +1,6 @@
 package models;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import org.hibernate.annotations.ColumnDefault;
 
 import javax.persistence.*;
 import java.util.Date;
